Rename some '.d' files
