+++ js/renovation/ui/pager/common/light_button.tsx
@@ -6,7 +6,7 @@ import { PAGER_CLASS } from './consts';
 import { closestClass } from '../utils/closest_class';
 import { subscribeToClickEvent } from '../../../utils/subscribe_to_event';
 import { DisposeEffectReturn } from '../../../utils/effect_return';
-import { EventCallback } from '../../common/event_callback.d';
+import { EventCallback } from '../../common/event_callback';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const viewFunction = ({
